Refactor and reorganize actions and reducers
