+++ client/src/rootReducer.js
@@ -1,15 +1,17 @@
 import { combineReducers } from 'redux';
 import { routerReducer } from 'react-router-redux';
 import entries from './components/entries/entries.reducer.js';
-import signUp from './components/landing/user_signup.reducer.js';
+import signup from './components/signup/signup.reducer.js';
 import user from './components/landing/user.reducer.js';
 import trends from './components/trends/trends.reducer.js';
+import calling from './components/entries/calling.reducer.js';
 
 const rootReducer = combineReducers({
   entries,
-  signUp,
+  signup,
   user,
   trends,
+  calling,
   router: routerReducer
 });
